Modify exercise - The Garden Application

+++ 09) Inheritance week-11/02) Inheritance/02) The Garden Application/Plant.cs
@@ -26,7 +26,7 @@ namespace _02__The_Garden_Application
             }
         }
         
-        public virtual void Feed(int waterInput)
+        public virtual void Feed(float waterInput)
         {
             WaterAmount += waterInput;
         }
